docs: split custom-collector guide into python and c variants
- Replace custom-collector.md with python-custom-collector.md and c-custom-collector.md
- Update README.md and monitor/README.md with links to both guides
- Register both guides in mkdocs.yml nav
