+++ jumper_extension/config/collectors/c/collectors.yaml
