+++ jumper_extension/monitor/pipeline.py
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
 
 from jumper_extension.adapters.data import NodeInfo
 from jumper_extension.config.utils import instantiate
-from jumper_extension.config.python import load_collectors_config
+from jumper_extension.config.collectors.python import load_collectors_config
 
 
 class PipelineBuilder:
