+++ mkdocs.yml
@@ -22,7 +22,8 @@ nav:
           - Python Script Export: guides/script-export.md
           - Plot Export: guides/plot-export.md
           - Custom Monitors: guides/custom-monitor.md
-          - Custom Collectors: guides/custom-collector.md
+          - Custom Python Collectors: guides/python-custom-collector.md
+          - Custom C Collectors: guides/c-custom-collector.md
           - Visualizing Custom Collector Metrics: guides/visualizing-custom-collector-metrics.md
       - Public API:
           - Overview: api/index.md
